+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Show paths to config files when configuration errors occur
 
+* `jj fix` now supports configuring the default revset for `-s` using the
+  `revsets.fix` config.
+
 ### Fixed bugs
 
 ## [0.18.0] - 2024-06-05